+++ actions/action-install-fluvio-cluster.sh
@@ -23,9 +23,5 @@ if [ "$CLUSTER_TYPE" = "local" ]; then
     LOCAL_FLAG="--local"
 fi
 
-# If VERSION is equal to exactly "latest", use LATEST channel
-if [ "${VERSION}" == "latest" ]; then
-    fluvio cluster start --rust-log $RUST_LOG  $LOCAL_FLAG --spu $SPU_NUMBER --chart-version="${CHART_VERSION}"
-else
-    fluvio cluster start --rust-log $RUST_LOG  $LOCAL_FLAG  --spu $SPU_NUMBER
-fi
+
+fluvio cluster start --rust-log $RUST_LOG  $LOCAL_FLAG --spu $SPU_NUMBER