+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix Replication bug in K8 ([#1290](https://github.com/infinyon/fluvio/pull/1290))
 * Add apply method to `StoreContext`. ([#1289](https://github.com/infinyon/fluvio/pull/1289))
 * Build M1 mac CLI ([#132](https://github.com/infinyon/fluvio/pull/1312))
+* Use inline helm chart ([#1292](https://github.com/infinyon/fluvio/pull/1292))
 
 ## Platform Version 0.8.5 - 2021-07-14
 * Add unstable Admin Watch API for topics, partitions, and SPUs ([#1136](https://github.com/infinyon/fluvio/pull/1136))